feat(): refactor the project structure

+++ node-blog-server/src/router/blog.js
@@ -1,13 +1,19 @@
 const { model } = require('m2-node');
-const { getList, getDetail, createItem, updateItem, deleteItem } = require('../controllers/blog');
+const { getList, getDetail, createItem, updateItem, deleteItem } = require('../controller/blog');
+const { checkIsAuth } = require('../util/auth');
 const { SuccessModel, ErrorModel } = model;
 
 const blogRouter = (req, res) => {
-  const { method, path, query, body } = req;
+  const { method, path, query, body, session } = req;
   const { id } = query;
 
   if (method === 'GET' && path === '/api/blog/list') {
-    const { author, keyword } = query;
+    let { author, keyword, isadmin } = query;
+    if (isadmin) {
+      const result = checkIsAuth(session);
+      if (result) return result;
+      author = session.data.username
+    }
     return getList(author, keyword).then(result => {
       return new SuccessModel(result);
     });
@@ -20,7 +26,10 @@ const blogRouter = (req, res) => {
   }
 
   if (method === 'POST' && path === '/api/blog/new') {
-    body.author = 'miracle';
+    const result = checkIsAuth(session);
+    if (result) return result;
+
+    body.author = session.data.username;
     return createItem(body).then(result => {
       if (!result.id) return new ErrorModel('添加博客失败');
       return new SuccessModel(result);
@@ -28,7 +37,10 @@ const blogRouter = (req, res) => {
   }
 
   if (method === 'POST' && path === '/api/blog/update') {
-    body.author = 'miracle';
+    const result = checkIsAuth(session);
+    if (result) return result;
+
+    body.author = session.data.username;
     return updateItem(id, body).then(result => {
       if (!result) return new ErrorModel('更新博客失败');
       return new SuccessModel();
@@ -36,7 +48,10 @@ const blogRouter = (req, res) => {
   }
 
   if (method === 'POST' && path === '/api/blog/del') {
-    body.author = 'miracle';
+    const result = checkIsAuth(session);
+    if (result) return result;
+
+    body.author = session.data.username;
     return deleteItem(id, body.author).then(result => {
       if (!result) return new ErrorModel('删除博客失败');
       return new SuccessModel();
